backend/requirements.txt
fastapi
uvicorn
python-multipart
passlib[bcrypt]
python-dotenv
supabase
implicit
numpy
pandas
scikit-learn
torch
torchvision
sentence-transformers
scipy
joblib
Pillow
matplotlib
tqdm
pydantic[email]
requests
python-multipart